Tweak how we do polyfilling to be less aggressive
Previously we were including the polyfill in the final transpiled lib. This caused
issues when requiring it as libraries that are requireing us may also want to
include the polyfill. Unfortuately, the polyfill will cause errors if it's included
more than once.

Longer term we need a better solution for this. In the shorter term the convention
is that the "top level" library should include the polyfilly when building. This
lib now only includes the polyfill when running it's own tests.